Kukushka wrapper fix for iOS WebKit approach
There was no message handling for WebKit approach which resulted onLoadFail error action fire due time out. It was incorrect cuz the page has successfully loaded and no errors have expected to be there.

I also set wkContentMode force to 0 cuz it provides best result. Feel free to test with it. wkAlloswLinkPreview is not nessesary but I also included it.